pyproject.toml
[project]
name = "munich-sushi-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "eventlet>=0.33.0",
    "faiss-cpu>=1.11.0",
    "fastmcp>=2.8.1",
    "flask>=3.0.0",
    "flask-socketio>=5.3.0",
    "google-api-python-client>=2.172.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.0",
    "google-maps-reviews>=0.0.3",
    "googlemaps>=4.10.0",
    "ipykernel>=6.29.5",
    "jq>=1.7",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.2",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.24",
    "langgraph>=0.4.8",
    "langgraph-cli[inmem]>=0.3.3",
    "langmem>=0.0.27",
    "ollama>=0.5.1",
    "protobuf>=3.20.0",
    "python-dotenv>=1.0.0",
    "python-engineio>=4.7.0",
    "python-socketio>=5.9.0",
    "requests>=2.32.4",
    "setuptools>=80.9.0",
    "streamlit>=1.46.0",
    "tavily-python>=0.7.7",
]
